updated shallow-water (MPI) makefile and removed compile-time errors in ConservativeShallowWaterClass_MPI.f90
Still need to clean up the ConservativeShallowWaterClass_MPI.f90 warnings and do a run-time test.
In order to do a runtime test, we need a parallel mesh. I'm working on a program to automate
the domain decomposition. The first approach will look at using an eigenvalue solver for the
mesh partitioning.