Some cleaning up was done....
For documentation, there will be a -TechnicalDocumentation and a -Manual.
The former will motivate the need for the item being documented, and will
describe the mathematics, algorithms, and will have convergence
studies, timing studies, memory use studies (profiling,etc). The latter
will provide documentation on the data-structures, subroutines and functions,
and how-to-use. After this commit, a branch will be started. The master branch
will trim out the highend solvers examples that have not been fully explored.